Use integer scaling for HpComponent

+++ ReadyM.Wukong.Common/ECS/Components/HpComponent.cs
@@ -11,5 +11,5 @@ public partial struct HpComponent : IOwnershipBased
     private float _hp;
     private float _hpMaxBase;
     private bool _isDead;
-    private float _hpMultiplier;
+    private int _hpMultiplier;
 }
